pull out suite running logic to new module
